add gRpc

+++ internal/dto/order_dto.go
@@ -1,15 +1,15 @@
 package dto
 
 type OrderInputDTO struct {
-	Price float64 `json:"price"`
-	Tax   float64 `json:"tax"`
+	Price float32 `json:"price"`
+	Tax   float32 `json:"tax"`
 }
 
 type OrderOutputDTO struct {
 	ID         string  `json:"id"`
-	Price      float64 `json:"price"`
-	Tax        float64 `json:"tax"`
-	FinalPrice float64 `json:"final_price"`
+	Price      float32 `json:"price"`
+	Tax        float32 `json:"tax"`
+	FinalPrice float32 `json:"final_price"`
 }
 
 type ListOrderOutputDTO struct {
